feat(console): user logs
