+++ packages/console/src/components/AuditLogTable/components/EventName/index.tsx
