+++ packages/console/src/components/AuditLogTable/components/EventSelector/index.tsx
